The routine for generating R safe column names is no longer necessary for Plotter because original sample names are now preserved in the DGE step. It has been moved to util.py for any potential future use, and to preserve the time spent.
